Fix two major bugs, save memory, etc
The two bugs are:
	Several backgrounds had a weird flashing palette animation, when
they're supposed to not have any palette animation at all.
	Interpolated backgrounds didn't look correct on certain surface sizes -
it's still not perfect on negative Y.

The thumbnail cache is now cleared when the settings menu is closed.

I changed the shader to pre-calculate some values instead of using
uniforms. It's probably pointless, but I think it's better this way.

I also switched to using the default 0-4 values from the ROM instead of
my remapped 0-2 values for animation types. I have no idea what the
difference between values 2 and 4 are, though.